pyproject.toml
[tool.poetry]
name = "geli"
version = "0.1.0"
description = ""
authors = ["Anirudh Jonnalagadda <anirudh@avigna.ai>",
           "Kartik Rathee <kartik.rathee@avigna.ai>",
          ]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
nrel-pysam = "5.1.0"
pandas = "2.2.2"
fastapi = "^0.111.0"
fastapi-healthz = "^0.5.1"
uvicorn = "0.15.0"
aiohttp = "^3.9.0"
boto3= "1.34.85"
pytest-mock = "^3.14.0"
aiosignal = "1.3.1"
annotated-types = "0.6.0"
anyio = "3.7.1"
asgiref = "3.8.1"
astor = "0.8.1"
asttokens = "2.4.1"
asyncio = "3.4.3"
attrs = "23.2.0"
black = "23.12.1"
botocore = "1.34.85"
build = "1.2.1"
cachecontrol = "0.14.0"
certifi = "2024.2.2"
charset-normalizer = "3.3.2"
cleo = "2.1.0"
click = "8.1.7"
colorama = "0.4.6"
crashtest = "0.4.1"
decorator = "5.1.1"
distlib = "0.3.8"
dulwich = "0.21.7"
exceptiongroup = "1.2.0"
executing = "2.0.1"
fastjsonschema = "2.19.1"
filelock = "3.13.3"
flake8 = "4.0.1"
flake8-bugbear = "22.12.6"
flake8-docstrings = "1.7.0"
flake8-polyfill = "1.0.2"
flake8-use-fstring = "1.4"
flynt = "0.77"
frozenlist = "1.4.1"
greenlet = "3.0.3"
h11 = "0.14.0"
idna = "3.7"
importlib-metadata = "7.1.0"
iniconfig = "2.0.0"
installer = "0.7.0"
ipython = "8.23.0"
isort = "5.13.2"
jaraco-classes = "3.4.0"
jedi = "0.19.1"
jmespath = "1.0.1"
keyring = "24.3.1"
matplotlib-inline = "0.1.7"
mccabe = "0.6.1"
more-itertools = "10.2.0"
msgpack = "1.0.8"
multidict = "6.0.5"
mypy-extensions = "1.0.0"
numpy = "1.26.4"
packaging = "24.0"
parso = "0.8.4"
pathlib = "1.0.1"
pathspec = "0.12.1"
pep8-naming = "0.12.1"
pexpect = "4.9.0"
platformdirs = "4.2.0"
pluggy = "1.4.0"
prompt-toolkit = "3.0.43"
ptyprocess = "0.7.0"
pure-eval = "0.2.2"
pycodestyle = "2.8.0"
pydantic = "^2.8.2"
pydantic-core = "^2.8.2"
pydocstyle = "6.3.0"
pyflakes = "2.4.0"
pygments = "2.17.2"
pyproject-hooks = "1.0.0"
python-dateutil = "2.9.0.post0"
python-dotenv = "1.0.1"
pytz = "2024.1"
pywin32-ctypes = "0.2.2"
rapidfuzz = "3.8.1"
requests = "2.31.0"
requests-toolbelt = "1.0.0"
s3transfer = "0.10.1"
shellingham = "1.5.4"
six = "1.16.0"
sniffio = "1.3.1"
snowballstemmer = "2.2.0"
sqlalchemy = "2.0.30"
stack-data = "0.6.3"
starlette = "^0.37.2"
tomli = "2.0.1"
tomlkit = "0.12.4"
traitlets = "5.14.2"
trove-classifiers = "2024.3.25"
typing-extensions = "4.12.2"
tzdata = "2024.1"
urllib3 = "2.2.1"
wcwidth = "0.2.13"
yarl = "1.9.4"
zipp = "3.18.1"
openpyxl = "^3.1.3"
psycopg2 = "^2.9.9"
pyjwt = "^2.8.0"
cryptography = "^42.0.8"
pynamodb = "^6.0.1"
aws-secretsmanager-caching = "^1.1.3"


[tool.poetry.dev-dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^4.0.1"
flake8-bugbear = "^22.12.6"
pep8-naming = "^0.12.1"
flake8-docstrings = "^1.7.0"
flake8-use-fstring = "^1.4"
flynt = "^0.77"
ipython = "^8.18.1"
coverage = "^7.5.1"
pytest-coverage = "^0.0"
jupyter = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py310']
preview = true
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 88

